Move Account, Payment, and User entities to their own folder
Move Account, Payment, User classes from /entity into the /domain/entity folder
